# internal/entities/inside.go
package entities

type Inside struct {
	Pinid string `json:"pinid,omitempty"`
	Polid string `json:"polid,omitempty"`
}

type Inside1 struct {
	Pinid string `json:"pinid,omitempty"`
}

type Inside2 struct {
	Polid string `json:"polid,omitempty"`
}